backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
yt-dlp>=2024.3.10
google-genai>=1.0.0
pydantic>=2.10.0
